First commit, working on flashcard ranking system

+++ flashcard.h
@@ -2,9 +2,9 @@
 
 typedef struct {
     int index;
-    int correct;
-    int attempts;
-    int familiarity;
+    float correct;
+    float attempts;
+    float familiarity;
     char spanish_word[20];
     char english_word[20];
 } Flashcard ;
