+++ servicer.h
@@ -12,11 +12,11 @@ int edit_english_word(char *en_edit, int index);
 
 int edit_spanish_word(char *es_edit, int index);
 
-int edit_familiarity_score(int fam_edit, int index);
+int edit_familiarity_score(float fam_edit, int index);
 
-int edit_total_attempts(int att_edit, int index);
+int edit_total_attempts(float att_edit, int index);
 
-int edit_correct_attempts(int cor_edit, int index);
+int edit_correct_attempts(float cor_edit, int index);
 
 int submit_flashcard_data(char *en_word, char *es_word);
